requirements.txt
fastapi
uvicorn
tortoise-orm
asyncpg
pydantic